+++ packages/vike-react/src/config.ts
@@ -108,11 +108,16 @@ const config = {
       cumulative: true,
       env: { client: true, server: true },
     },
+    // TODO/next-major: move to +react.js > strictMode ?
     reactStrictMode: {
       env: { client: true, server: true },
     },
     Loading: {
       env: { server: true, client: true },
+    },
+    react: {
+      cumulative: true,
+      env: {},
     },
   },
 } satisfies Config