+++ packages/vike-react/src/index.ts
@@ -1,4 +1,4 @@
-// TODO/next-major-release: remove this file/export
+// TODO/next-major: remove this file/export
 console.warn(
   "[vike-react][warning][deprecation] Replace `import vikeReact from 'vike-react'` with `import vikeReact from 'vike-react/config'` (typically in your /pages/+config.js)",
 )