+++ packages/vike-react/CHANGELOG.md
@@ -1,3 +1,40 @@
+## [0.6.1](https://github.com/vikejs/vike-react/compare/vike-react@0.6.0...vike-react@0.6.1) (2025-03-16)
+
+
+### Bug Fixes
+
+* allow user to override pageContext.{pageHtmlString,pageHtmlStream} (vikejs/vike[#389](https://github.com/vikejs/vike-react/issues/389), https://github.com/vikejs/vike/issues/389#issuecomment-2725260763) ([95cf570](https://github.com/vikejs/vike-react/commit/95cf5706c962d6ed6a13860630d9fc65275a2a2b)), closes [/github.com/vikejs/vike/issues/389#issuecomment-2725260763](https://github.com//github.com/vikejs/vike/issues/389/issues/issuecomment-2725260763)
+
+
+
+# [0.6.0](https://github.com/vikejs/vike-react/compare/vike-react@0.5.13...vike-react@0.6.0) (2025-03-11)
+
+
+### Bug Fixes
+
+* update react-streaming to `^0.4.2` ([#168](https://github.com/vikejs/vike-react/issues/168)) ([9ae0bfb](https://github.com/vikejs/vike-react/commit/9ae0bfbc974be8421a672ed485ebba484a8c1153))
+
+
+### BREAKING CHANGES
+
+* Update React to `19.0.0` or above.
+
+
+
+## [0.5.13](https://github.com/vikejs/vike-react/compare/vike-react@0.5.12...vike-react@0.5.13) (2025-02-16)
+
+
+### Bug Fixes
+
+* support `stream: false` ([702d05e](https://github.com/vikejs/vike-react/commit/702d05ee163a981f882fec88c93491b3c6e4302e))
+
+
+### Features
+
+* +react.{client,server}.js ([#165](https://github.com/vikejs/vike-react/issues/165)) ([831e451](https://github.com/vikejs/vike-react/commit/831e451aadf5b2798ec2ffc9c310562ec6184dff))
+
+
+
 ## [0.5.12](https://github.com/vikejs/vike-react/compare/vike-react@0.5.11...vike-react@0.5.12) (2025-01-22)